refactor: extract decompression logic to shared utility
- Created FileDecompressor utility class in Core/Utilities
- Eliminated code duplication between ImportService and ImportDialog
- Single source of truth for .gz file decompression
- Easier to maintain and test

This addresses the recurring code review feedback about duplication.
